Rework the way that binding works
Fields have to be declared as either Optional or Required when they are defined. (Trying to convert between optional and required led to a big mess.)

Some cleanup is definitely needed. Right now there are some intermediate objects between a RequestKey and a RequestBindingWithRequest that might not need to exist. Also need to add some friendly and concise ways of declaring fields.